Adjust controller behavior when being uninstalled
A new "controller" watches the Deployment for an uninstall annotation.
When the annotation is true, it updates a flag so that all the other
controllers don't have to track it individually. Most controllers just
skip their reconciles during the uninstall, but the template-sync will
clean up its finalizers (and skip an error case during that process).

Refs:
 - https://issues.redhat.com/browse/ACM-3329

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>